data_provider/requirements.txt
dateparser
feedparser
goose3
jsonlines
loguru
newscatcherapi
newspaper3k
typer
